Orgainised the files into packjages

+++ WebApplication2/WebApplication2/Data Layer/OrderMapper.cs
@@ -140,7 +140,7 @@ namespace WebApplication2
             return "Edit Order COmpleted";
         }
 
-        public static string AddProduct(Order ord, int pID, int amnt)
+        public static string AddProduct(Order ord, int pID, double amnt)
         {
             SqlConnection connection = new SqlConnection(@"Data Source=(LocalDB)\v11.0;AttachDbFilename=C:\Users\werl\Documents\Visual Studio 2013\Projects\SystemsAnalysis\WebApplication2\WebApplication2\App_Data\Database.mdf;Integrated Security=True");
